refactor(dev-wallet): fix lint errors
