+++ packages/apps/dev-wallet/src/pages/backup-recovery-phrase/write-down/confirm-recovery-phrase.tsx
@@ -1,4 +1,4 @@
-import { useWallet } from '@/hooks/wallet.context';
+import { useWallet } from '@/hooks/wallet.hook';
 import { Box, Button, Heading, Text, TextField } from '@kadena/react-ui';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
