+++ packages/apps/dev-wallet/src/pages/unlock-wallet/unlock-wallet.tsx
@@ -2,7 +2,7 @@ import { Box, Button, Heading, Text, TextField } from '@kadena/react-ui';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { Link, Navigate, useParams } from 'react-router-dom';
-import { useWallet } from '../../hooks/wallet.context';
+import { useWallet } from '../../hooks/wallet.hook';
 
 export function UnlockWallet() {
   const { register, handleSubmit } = useForm<{ password: string }>();
