+++ packages/apps/dev-wallet/src/utils/chrome.d.ts
@@ -1,0 +1,12 @@
+interface Window {
+  chrome?: {
+    extension?: {
+      getBackgroundPage?: () => Window;
+    };
+    runtime?: {
+      onMessage?: {
+        addListener?: (listener: (message: string) => void) => void;
+      };
+    };
+  };
+}
