+++ packages/apps/dev-wallet/src/utils/window.ts
@@ -1,27 +1,28 @@
+// <reference types="chrome"/>
 /**
  * Detect what context the app is running in.
  *
  * Mainly used for detecting extension context which needs a different routing strategy
  */
 export const getScriptType = () => {
-  const { chrome } = window as any;
+  const { chrome } = window;
   if (
     chrome &&
     chrome.extension &&
     chrome.extension.getBackgroundPage &&
     chrome.extension.getBackgroundPage() === window
   ) {
-    return "BACKGROUND";
+    return 'BACKGROUND';
   } else if (
     chrome &&
     chrome.extension &&
     chrome.extension.getBackgroundPage &&
     chrome.extension.getBackgroundPage() !== window
   ) {
-    return "POPUP";
+    return 'POPUP';
   } else if (!chrome || !chrome.runtime || !chrome.runtime.onMessage) {
-    return "WEB";
+    return 'WEB';
   } else {
-    return "CONTENT";
+    return 'CONTENT';
   }
 };
